+++ rst/conf.py
@@ -27,7 +27,8 @@ extensions = ['sphinx.ext.autodoc', 'sphinx.ext.intersphinx',
               'sphinx_cython', 'sphinxcontrib.asyncio' ]
 
 # Link to Python standard library
-intersphinx_mapping = {'python': ('https://docs.python.org/3/', None) }
+intersphinx_mapping = {'python': ('https://docs.python.org/3/', None),
+                       'trio': ('https://trio.readthedocs.io/en/stable/', None),}
 
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ['_templates']